postponing choice on lonlatdepth_dtype to ParticleSet.__init__()

+++ tests/test_particle_sets.py
@@ -32,8 +32,8 @@ def test_pset_create_lon_lat(fieldset, mode, npart=100):
 @pytest.mark.parametrize('mode', ['scipy', 'jit'])
 @pytest.mark.parametrize('lonlatdepth_dtype', [np.float64, np.float32])
 def test_pset_create_line(fieldset, mode, lonlatdepth_dtype, npart=100):
-    lon = np.linspace(0, 1, npart, dtype=np.float32)
-    lat = np.linspace(1, 0, npart, dtype=np.float32)
+    lon = np.linspace(0, 1, npart, dtype=lonlatdepth_dtype)
+    lat = np.linspace(1, 0, npart, dtype=lonlatdepth_dtype)
     pset = ParticleSet.from_line(fieldset, size=npart, start=(0, 1), finish=(1, 0),
                                  pclass=ptype[mode], lonlatdepth_dtype=lonlatdepth_dtype)
     assert np.allclose([p.lon for p in pset], lon, rtol=1e-12)
